+++ models/feedback_model.go
@@ -8,7 +8,7 @@ import (
 
 type Feedback struct {
 	ID        uuid.UUID `gorm:"type:uuid;default:uuid_generate_v4();primary_key" json:"id"`
-	Type      int       `json:"type"`
+	Type      int8      `json:"type"`
 	UserID    uuid.UUID `gorm:"type:uuid;not null" json:"userID"`
 	User      User      `json:"user"`
 	Content   string    `json:"content"`
