+++ schemas/org_schemas.go
@@ -7,6 +7,6 @@ used in controllers/organization_controllers/reviews.go
 */
 type ReviewCreateSchema struct {
 	Content   string `json:"content" validate:"required,max=500"`
-	Rating    int    `json:"rating"  validate:"required,min=1,max=5"`
+	Rating    int8   `json:"rating"  validate:"required,min=1,max=5"`
 	Anonymous bool   `json:"isAnonymous"`
 }
